refactor:  reorganize gem analyzer
